refactor: Implement just-in-time login in the booking view by extracting the login form into a reusable component and integrating it into both the booking and login views.
