+++ src/components/auth/LoginForm.vue
@@ -1,0 +1,174 @@
+<script setup lang="ts">
+import { ref } from 'vue'
+import { useAuthStore } from '../../stores/useAuthStore'
+import { useRouter } from 'vue-router'
+import { useI18n } from 'vue-i18n'
+
+const props = defineProps<{
+  redirect?: string
+  embedded?: boolean
+}>()
+
+const emit = defineEmits<{
+  (e: 'success'): void
+}>()
+
+const authStore = useAuthStore()
+const router = useRouter()
+useI18n()
+
+const email = ref('')
+const otpCode = ref('')
+const codeSent = ref(false)
+
+async function sendCode() {
+  if (!email.value) return
+  
+  try {
+    await authStore.sendOtpCode(email.value)
+    codeSent.value = true
+  } catch (error) {
+    console.error('Failed to send code:', error)
+  }
+}
+
+async function verifyCode() {
+  if (!email.value || !otpCode.value) return
+  
+  try {
+    await authStore.verifyOtpCode(email.value, otpCode.value)
+    
+    // Provide a small delay to ensure state updates propagate
+    // Checks if redirect prop is present
+    if (props.redirect) {
+         router.push(props.redirect)
+    } else {
+        // Emit success for parent to handle (e.g., embedded in booking)
+        emit('success')
+    }
+  } catch (error) {
+    console.error('Verification failed:', error)
+  }
+}
+
+function goBack() {
+  codeSent.value = false
+  otpCode.value = ''
+}
+</script>
+
+<template>
+  <div :class="embedded ? 'w-full' : 'max-w-md w-full'">
+    <!-- Login Card -->
+    <div :class="[
+      'bg-white rounded-lg',
+      embedded ? '' : 'border border-gray-200 shadow-sm p-8'
+    ]">
+      <!-- Step 1: Email Input -->
+      <div v-if="!codeSent">
+        <h2 v-if="!embedded" class="text-2xl font-bold text-gray-900 mb-6 text-center">{{ $t('auth.login') }}</h2>
+        <h2 v-else class="text-xl font-bold text-gray-900 mb-4">{{ $t('auth.login_to_continue') }}</h2>
+
+        <!-- Error Message -->
+        <div v-if="authStore.error" class="bg-red-50 border border-red-200 text-red-800 px-4 py-3 rounded-lg mb-6">
+          {{ authStore.error }}
+        </div>
+
+        <form @submit.prevent="sendCode" class="space-y-6">
+          <div>
+            <div class="relative">
+              <input
+                v-model="email"
+                type="email"
+                required
+                :placeholder="$t('auth.email')"
+                class="w-full px-4 py-3 border border-gray-300 rounded-md focus:ring-1 focus:ring-primary-500 focus:border-primary-500 placeholder-gray-400"
+                :disabled="authStore.loading"
+              />
+              <div class="absolute inset-y-0 right-0 pr-3 flex items-center pointer-events-none">
+                <svg class="h-5 w-5 text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+                  <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M3 8l7.89 5.26a2 2 0 002.22 0L21 8M5 19h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v10a2 2 0 002 2z"></path>
+                </svg>
+              </div>
+            </div>
+          </div>
+
+          <button
+            type="submit"
+            :disabled="authStore.loading || !email"
+            class="w-full flex items-center justify-center gap-2 bg-primary-600 hover:bg-primary-700 text-white font-semibold py-3 px-4 rounded-md transition-all disabled:opacity-50 disabled:cursor-not-allowed shadow-sm"
+          >
+            <div v-if="authStore.loading" class="h-5 w-5 border-2 border-white border-t-transparent rounded-full animate-spin"></div>
+            <span>{{ authStore.loading ? $t('common.sending') : $t('auth.send_code') }}</span>
+          </button>
+        </form>
+
+        <!-- Divider -->
+        <div class="relative my-8">
+          <div class="absolute inset-0 flex items-center">
+            <div class="w-full border-t border-gray-300"></div>
+          </div>
+          <div class="relative flex justify-center text-sm">
+            <span class="px-4 bg-white text-gray-500">{{ $t('auth.verification_email_notice') }}</span>
+          </div>
+        </div>
+      </div>
+
+      <!-- Step 2: Code Input -->
+      <div v-else>
+        <button
+          @click="goBack"
+          class="text-primary-600 hover:text-primary-700 font-medium mb-4 flex items-center text-sm"
+        >
+          ← {{ $t('auth.change_email') }}
+        </button>
+
+        <h2 v-if="!embedded" class="text-2xl font-bold text-gray-900 mb-2 text-center">{{ $t('auth.enter_code') }}</h2>
+        <h2 v-else class="text-xl font-bold text-gray-900 mb-2">{{ $t('auth.enter_code') }}</h2>
+        
+        <p class="text-gray-600 mb-6" :class="embedded ? '' : 'text-center'" v-html="$t('auth.code_sent_to', { email: `<span class='font-semibold'>${email}</span>` })"></p>
+
+        <!-- Error Message -->
+        <div v-if="authStore.error" class="bg-red-50 border border-red-200 text-red-800 px-4 py-3 rounded-lg mb-6">
+          {{ authStore.error }}
+        </div>
+
+        <form @submit.prevent="verifyCode" class="space-y-6">
+          <div>
+            <input
+              v-model="otpCode"
+              type="text"
+              required
+              :placeholder="$t('auth.enter_6_digit')"
+              maxlength="6"
+              pattern="[0-9]{6}"
+              class="w-full px-4 py-3 border border-gray-300 rounded-md focus:ring-1 focus:ring-primary-500 focus:border-primary-500 placeholder-gray-400 text-center text-2xl tracking-widest font-mono"
+              :disabled="authStore.loading"
+              autofocus
+            />
+          </div>
+
+          <button
+            type="submit"
+            :disabled="authStore.loading || otpCode.length !== 6"
+            class="w-full flex items-center justify-center gap-2 bg-primary-600 hover:bg-primary-700 text-white font-semibold py-3 px-4 rounded-md transition-all disabled:opacity-50 disabled:cursor-not-allowed shadow-sm"
+          >
+            <div v-if="authStore.loading" class="h-5 w-5 border-2 border-white border-t-transparent rounded-full animate-spin"></div>
+            <span>{{ authStore.loading ? $t('common.verifying') : $t('auth.verify_code') }}</span>
+          </button>
+        </form>
+
+        <!-- Resend Code -->
+        <div class="mt-6 text-center">
+          <button
+            @click="sendCode"
+            :disabled="authStore.loading"
+            class="text-sm text-primary-600 hover:text-primary-700 font-medium disabled:opacity-50"
+          >
+            {{ $t('auth.resend_code') }}
+          </button>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
